+++ src/app/board/page.tsx
@@ -1,11 +1,6 @@
-'use client';
 import AllPage from './all/page';
 
 const BoardPage = () => {
-  // useEffect(() => {
-  //   throw new Error('테스트용 에러 발생!');
-  // }, []);
-
   return <AllPage />;
 };
